requirements.txt
django
psycopg
load_dotenv